created a medical quiz of 5 questions, after the answers it shows how many correct answers from 5 questions
