+++ dwin_quiz/src/component/quiz-pages/medical-quiz-page/MedicalQuizPage.js
@@ -1,10 +1,7 @@
-import React, { useState } from "react"
+import React from 'react'
+import Quiz from '../Quiz'
 import './MedicalQuizPage.scss'
-import Medical from '../../../img/Medical.jpg'
-import Timer from "../../timer/Timer"
-import Medal from '../../../img/medal-1.svg'
-import { useNavigate } from 'react-router';
-
+import MedicalImg from '../../../img/Medical.jpg'
 
 
 export default function MedicalQuiz() {
@@ -56,83 +53,11 @@ export default function MedicalQuiz() {
         },
     ]
 
-    const [currentQuestion, setCurrentQuestion] = useState(0)
-    const [showScore, setShowScore] = useState(false)
-    const [isCorrectVal, setIsCorrectVal] = useState(0)
-
-
-    const handledAnswerCheckBoxClick = (val) => {
-
-        if (val === true) {
-            setIsCorrectVal(isCorrectVal + 1)
-        }
-
-        const nextCurrent = currentQuestion + 1
-        if (nextCurrent === questions.length) {
-            setShowScore(true)
-        }
-        setCurrentQuestion(nextCurrent)
-    }
-    const navigate = useNavigate();
-
+    
     return (
 
         <div className='medicalQuiz'>
-            {showScore ? (
-                <>
-                    <div className='score-section'>
-                        You scored {isCorrectVal} out of {questions.length}
-
-                        {isCorrectVal >= 3 &&
-                            <div className='medalImg'>
-                                <img src={Medal} />
-                                <h4>Good Job</h4>
-                            </div>}
-
-                        <button
-                            className='buttonGoNewQuiz'
-                            onClick={() => navigate('/selectTopic')}>
-                            Go New Quiz</button>
-
-                    </div>
-                </>
-            ) : (
-                <>
-                    <div className='title'>
-
-                        <div>
-                            <h1>Medical  Quiz</h1>
-                            <h4> Read the following instructions </h4>
-                        </div>
-                        <div className='timer'>
-                            <Timer />
-                        </div>
-                    </div>
-                    <div className='bodyMedicalQuiz'>
-                        <div className='imgQuiz'>
-                            <img src={Medical} alt='imgMedicalQuiz' />
-                        </div>
-                        <div className='quiz'>
-
-                            <h2>Questions {currentQuestion + 1}/{questions.length}</h2>
-
-                            <div>
-                                {questions[currentQuestion].questionsText}
-                            </div>
-
-                        </div>
-
-                    </div>
-
-                    <div className='answer'>
-                        {questions[currentQuestion].answerOption.map((answerOption, i) =>
-                            <button onClick={() => handledAnswerCheckBoxClick(answerOption.isCorrect)} key={i}>{answerOption.answerText}</button>
-
-                        )}
-                    </div>
-                </>
-            )}
-
+           <Quiz title ={'Medical Quiz'} questions = {questions} img = {MedicalImg} /> 
         </div>
     )
 }
